unlink re_eval code blocks from op list
In the list of ops generated by something like /abc(?{...})def/,

	const(abc)
	null/special
	    ...
	const(...)
	const(def)

link the list, but skip the DO blocks. This means that for the runtime
case, we no longer need the temporary measure of deleting the DO blocks,
and it will facilitate the next step of handling literal code at runtime,
i.e. /$runtime(?{...})/.