FIXME this can't go into blead as is
This implements detection of a basic Schwartzian transform-style sort
block and removes it from the OP tree. It sets a new flag on the sort OP
which will indicate that this sort needs to be performed by one of the
special-case optimized C functions in pp_sort. The actual implementation
of that function is pending.