mktables: Verify input files' version
Since Unicode 3.2, all Unicode database source files (except
unfortunately, the most important one, UCD.txt), have their first lines
be identifying information, their name and version number.  This commit
checks that the version is the expected one.  This should prevent the
database from being out-of-sync.  Perl changes the names of some files
so that they are distinct on DOS filesystems, so we can't easily check
that the name in the file is the same as the name of the file.